Fix compiler warning
```
yocton_test.c: In function ‘run_test_with_limit’:
yocton_test.c:480:39: warning: format ‘%d’ expects argument of type ‘int’, but argument 4 has type ‘size_t’ {aka ‘long unsigned int’} [-Wformat=]
  480 |                 fprintf(stderr, "%s: %d bytes still allocated after test\n",
      |                                      ~^
      |                                       |
      |                                       int
      |                                      %ld
  481 |                         filename, alloc_test_get_allocated());
      |                                   ~~~~~~~~~~~~~~~~~~~~~~~~~~
      |                                   |
      |                                   size_t {aka long unsigned int}
```

+++ yocton_test.c
@@ -477,8 +477,8 @@ int run_test_with_limit(char *filename, int alloc_limit)
 	free(output);
 
 	if (alloc_test_get_allocated() != 0) {
-		fprintf(stderr, "%s: %d bytes still allocated after test\n",
-		        filename, alloc_test_get_allocated());
+		fprintf(stderr, "%s: %ld bytes still allocated after test\n",
+		        filename, (long) alloc_test_get_allocated());
 		success = 0;
 	}
